+++ frontend/src/components/Modal.tsx
@@ -1,5 +1,4 @@
-import { Typography, styled } from "@mui/material";
-import logo from "../assets/logo-1.png";
+import { styled } from "@mui/material";
 const StyledOverlay = styled("div")({
   width: "100%",
   height: "100%",
@@ -14,21 +13,8 @@ const StyledOverlay = styled("div")({
 });
 
 const StyledModalBody = styled("div")({
-  display: "flex",
-  flexDirection: "column",
-  justifyContent: "center",
-  alignItems: "center",
-  padding: "100px",
   backgroundColor: "#FFFFFF",
   borderRadius: "15px",
-});
-
-const StyledLogoContainer = styled("div")({
-  display: "flex",
-  paddingBottom: "20px",
-  justifyItems: "flex-start",
-  alignItems: "baseline",
-  width: "100%",
 });
 
 interface ModalProps {
@@ -42,15 +28,6 @@ const Modal: React.FC<ModalProps> = ({ children, onClose, isOpen }) => {
   return (
     <StyledOverlay onClick={onClose}>
       <StyledModalBody onClick={(e) => e.stopPropagation()}>
-        <StyledLogoContainer>
-          <img src={logo} alt="Logo" height="40" />
-          <Typography
-            variant="h4"
-            sx={{ color: "#004038", paddingLeft: "5px" }}
-          >
-            TaskTree
-          </Typography>
-        </StyledLogoContainer>
         {children}
       </StyledModalBody>
     </StyledOverlay>
